requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
passlib[bcrypt]
python-jose
bcrypt
requests
google-genai